requirements.txt
discord.py
prsaw
discord-py-slash-command
DiscordUtils
aiosqlite
topggpy
discord_components
youtube_dl
PyNaCl
giphy_client
aiofiles
discord_components